Long LD Register Accesses
Added register names to the places where the game uses the long form of the LD opcode (eg. ld a, [$ff4a]), which the disassembler ignored.

Also, replaced all of the EQUs in hardware.inc with equals signs after noticing that it was inadvertently using "EQU" in the comments as well (indicating a sloppy search/replace was done on that file by a previous steward of it)